+++ contracts/mixins/ERC223ReceiverMixin.sol
@@ -2,5 +2,5 @@ pragma solidity ^0.4.24;
 
 
 contract ERC223ReceiverMixin {
-  function tokenFallback(address _from, uint _value, bytes _data) public;
+  function tokenFallback(address _from, uint256 _value, bytes _data) public;
 }
