+++ contracts/mock/ERC223ReceiverMock.sol
@@ -3,9 +3,9 @@ import "../mixins/ERC223ReceiverMixin.sol";
 
 
 contract ERC223ReceiverMock is ERC223ReceiverMixin {
-  event Fallback(address indexed _from, uint _value, bytes _data);
+  event Fallback(address indexed _from, uint256 _value, bytes _data);
 
-  function tokenFallback(address _from, uint _value, bytes _data) public {
+  function tokenFallback(address _from, uint256 _value, bytes _data) public {
     emit Fallback(_from, _value, _data);
   }
 }
